ai-engine/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-chroma
chromadb
python-dotenv
pydantic